Fix google chart plugin configuration persistence
Nowadays, the chart wrapper returned by the editor seems not to be the same that it was given. For this plugin, this resulted in 'Chart Config' always starting from an initial configuration instead of the last edited state. 

This fixes the problem by saving the editor-returned chartWrapper variable for later reuse.